Break up Database impl into multiple files, organized by topic
